router-api/requirements.txt
fastapi==0.115.0
uvicorn==0.30.6
httpx==0.27.2
python-dotenv==1.0.1
pydantic==2.9.2
boto3==1.35.0
transformers==4.46.0
protobuf>=3.20.0